Split protein inference and scoring steps
