+++ assets/example_config/defaultSearchParams.json
@@ -141,13 +141,13 @@
             "max_diff": 0.01
         },
         "machine_learning": {
-            "max_psm_memory_mb": 2000,
+            "max_in_memory_table_mb": 2000,
             "force_oom": false,
             "min_trace_prob": 0.75,
             "max_q_value_mbr_itr": 0.20,
             "min_PEP_neg_threshold_itr": 0.90,
             "spline_points": 500,
-            "interpolation_points": 10,
+            "q_value_interpolation_points_per_bin": 10,
             "n_quantile_bins": 25,
             "enable_model_comparison": true,
             "validation_split_ratio": 0.2,
@@ -157,11 +157,14 @@
             "max_mbr_training_candidates": 1000000
         }
     },
-    "proteinInference": {
-        "min_peptides": 1
+    "proteinScoring": {
+        "min_peptides": 1,
+        "write_qc_plots": false,
+        "log_feature_importance": false
     },
     "maxLFQ": {
         "run_to_run_normalization": false,
+        "min_peptides": 1,
         "max_chunk_size_mb": 1024
     },
     "output": {
