+++ data/ecoli_test/ecoli_test_params.json
@@ -139,24 +139,26 @@
             "max_diff": 0.01
         },
         "machine_learning": {
-            "max_psm_memory_mb": 4000,
+            "max_in_memory_table_mb": 4000,
             "min_trace_prob": 0.75,
             "max_q_value_mbr_itr": 0.20,
             "min_PEP_neg_threshold_itr": 0.90,
             "spline_points": 500,
-            "interpolation_points": 10,
+            "q_value_interpolation_points_per_bin": 10,
             "enable_model_comparison": true,
             "validation_split_ratio": 0.2,
-            "qvalue_threshold": 0.01,
             "min_psms_for_comparison": 1000,
             "max_psms_for_comparison": 100000
         }
     },
-    "proteinInference": {
-        "min_peptides": 1
+    "proteinScoring": {
+        "min_peptides": 1,
+        "write_qc_plots": false,
+        "log_feature_importance": false
     },
     "maxLFQ": {
-        "run_to_run_normalization": false
+        "run_to_run_normalization": false,
+        "min_peptides": 1
     },
     "output": {
         "write_csv": true,
