+++ data/precompile/search_ecoli_prosit.json
@@ -87,16 +87,16 @@
             "max_diff": 0.01
         },
         "machine_learning": {
-            "max_psm_memory_mb": 400,
+            "max_in_memory_table_mb": 400,
             "min_trace_prob": 0.75,
             "max_q_value_mbr_itr": 0.20,
             "min_PEP_neg_threshold_itr": 0.90,
             "spline_points": 500,
-            "interpolation_points": 10
+            "q_value_interpolation_points_per_bin": 10
         }
     },
-    "proteinInference": {"min_peptides": 1},
-    "maxLFQ": {"run_to_run_normalization": false},
+    "proteinScoring": {"min_peptides": 1, "write_qc_plots": false, "log_feature_importance": false},
+    "maxLFQ": {"run_to_run_normalization": false, "min_peptides": 1},
     "output": {"write_csv": true, "write_decoys": true, "delete_temp": true, "plots_per_page": 12},
     "paths": {
         "library": "./data/precompile/ecoli_small_prosit.poin",
