+++ data/precompile/search_yeast_altimeter_OOM.json
@@ -87,16 +87,16 @@
             "max_diff": 0.01
         },
         "machine_learning": {
-            "max_psm_memory_mb": 0.01,
+            "max_in_memory_table_mb": 0.01,
             "min_trace_prob": 0.75,
             "max_q_value_mbr_itr": 0.20,
             "min_PEP_neg_threshold_itr": 0.90,
             "spline_points": 500,
-            "interpolation_points": 10
+            "q_value_interpolation_points_per_bin": 10
         }
     },
-    "proteinInference": {"min_peptides": 1},
-    "maxLFQ": {"run_to_run_normalization": true},
+    "proteinScoring": {"min_peptides": 1, "write_qc_plots": false, "log_feature_importance": false},
+    "maxLFQ": {"run_to_run_normalization": true, "min_peptides": 1},
     "output": {"write_csv": false, "write_decoys": false, "delete_temp": true, "plots_per_page": 12},
     "paths": {
         "library": "./temp/zenodo/yeast.poin",
