+++ src/Routines/SearchDIA/SearchMethods/PrecursorScoringSearch/README.md
@@ -1,10 +1,10 @@
-# ScoringSearch Module
+# PrecursorScoringSearch Module
 
-ScoringSearch is the 7th stage in the Pioneer DIA search pipeline, responsible for machine learning-based PSM rescoring, FDR control, and protein group analysis. This module features adaptive model selection and supports both in-memory and out-of-memory processing for datasets of varying sizes.
+PrecursorScoringSearch is the 7th stage in the Pioneer DIA search pipeline, responsible for machine learning-based PSM rescoring, FDR control, and protein group analysis. This module features adaptive model selection and supports both in-memory and out-of-memory processing for datasets of varying sizes.
 
 ## Overview
 
-ScoringSearch performs three main functions:
+PrecursorScoringSearch performs three main functions:
 1. **PSM Scoring**: Machine learning models (LightGBM or Probit Regression) rescore PSMs
 2. **FDR Control**: Calculate q-values and filter PSMs based on false discovery rate thresholds
 3. **Protein Inference**: Group peptides into minimal protein sets and calculate protein-level scores
@@ -13,13 +13,13 @@ ScoringSearch performs three main functions:
 
 ```
 ┌─────────────────┐    ┌──────────────────┐    ┌─────────────────────┐
-│   SecondPass    │───▶│   ScoringSearch  │───▶│ IntegrateChromato-  │
+│   SecondPass    │───▶│   PrecursorScoringSearch  │───▶│ IntegrateChromato-  │
 │   PSM Files     │    │                  │    │ gramSearch          │
 └─────────────────┘    └──────────────────┘    └─────────────────────┘
                               │
                               ▼
 ┌─────────────────────────────────────────────────────────────────────┐
-│                    ScoringSearch Pipeline                          │
+│                    PrecursorScoringSearch Pipeline                          │
 │                                                                     │
 │  ┌─────────────────┐  ┌──────────────────┐  ┌──────────────────┐   │
 │  │  1. PSM Scoring │─▶│ 2. FDR Control   │─▶│ 3. Protein       │   │
@@ -31,7 +31,7 @@ ScoringSearch performs three main functions:
 
 ## Model Selection Strategy
 
-ScoringSearch uses different strategies based on dataset size:
+PrecursorScoringSearch uses different strategies based on dataset size:
 
 ```
 PSM Count Decision Tree:
@@ -62,12 +62,12 @@ PSM Count Decision Tree:
 ### Memory Strategy
 
 ```
-                    PSM Count vs Memory Limit
+                    PSM Count vs Derived Row Budget
                               │
                  ┌────────────┼────────────┐
                  │            │            │
-        < max_psms_in_memory  │     ≥ max_psms_in_memory
-              (100K)          │           (100K)
+   < in-memory row budget     │    ≥ in-memory row budget
+  (from max_in_memory_table_mb)│   (from max_in_memory_table_mb)
                  │            │            │
                  ▼            │            ▼
           In-Memory Processing│    Out-of-Memory Processing
@@ -107,7 +107,7 @@ score_precursor_isotope_traces(
     max_q_value_lightgbm_rescore,
     max_q_value_mbr_itr,
     min_PEP_neg_threshold_itr,
-    max_psms_in_memory,
+    derived_max_in_memory_rows,
     q_value_threshold
 )
 ```
@@ -117,7 +117,7 @@ score_precursor_isotope_traces(
 ```
 score_precursor_isotope_traces()
 ├── get_psms_count(file_paths)
-├── if psms_count >= max_psms_in_memory:
+├── if psms_count >= derived_max_in_memory_rows:
 │   ├── sample_psms_for_lightgbm()
 │   ├── create_default_advanced_lightgbm_config()
 │   └── score_precursor_isotope_traces_out_of_memory!()
@@ -217,7 +217,7 @@ probit_regression_scoring_cv!()
 | **MBR Handling** | Separate MBR iterations | Creates compatibility columns |
 | **Hyperparameters** | Many (depth, eta, etc.) | Few (max_iter) |
 
-## Integration with ScoringSearch Pipeline
+## Integration with PrecursorScoringSearch Pipeline
 
 ### PSM Processing Output
 
@@ -231,9 +231,9 @@ psms[!, :MBR_is_best_decoy]            # MBR transfer metadata
 psms[!, :MBR_transfer_candidate]       # MBR filtering flags
 ```
 
-### 23-Step ScoringSearch Pipeline Integration
+### 23-Step PrecursorScoringSearch Pipeline Integration
 
-The trained models integrate into the broader ScoringSearch pipeline:
+The trained models integrate into the broader PrecursorScoringSearch pipeline:
 
 1. **Steps 1-3**: Model training and PSM scoring (this README's focus)
 2. **Steps 4-10**: PSM processing, merging, and FDR control
@@ -247,14 +247,15 @@ The trained models integrate into the broader ScoringSearch pipeline:
 {
     "optimization": {
         "machine_learning": {
-            "max_psms_in_memory": 100000,
+            "max_in_memory_table_mb": 2000,
             "enable_model_comparison": true,
             "max_q_value_lightgbm_rescore": 0.01,
             "max_q_value_mbr_itr": 0.20,
-            "min_PEP_neg_threshold_itr": 0.90
+            "min_PEP_neg_threshold_itr": 0.90,
+            "q_value_interpolation_points_per_bin": 10
         }
     },
-    "global_settings": {
+    "global": {
         "scoring": {
             "q_value_threshold": 0.01
         }
@@ -264,7 +265,7 @@ The trained models integrate into the broader ScoringSearch pipeline:
 
 ### Parameter Effects
 
-- **max_psms_in_memory**: Determines in-memory vs out-of-memory processing
+- **max_in_memory_table_mb**: Determines the derived in-memory row budget used to choose between in-memory and out-of-memory processing
 - **enable_model_comparison**: Controls automatic model selection for medium datasets
 - **max_q_value_mbr_itr**: Caps MBR transfers admitted to the iterative training (ITR) stage
 - **min_PEP_neg_threshold_itr**: Sets the PEP cutoff for relabeling weak targets as negatives during ITR
@@ -290,14 +291,12 @@ The trained models integrate into the broader ScoringSearch pipeline:
 ## Files and Structure
 
 ```
-ScoringSearch/
+PrecursorScoringSearch/
 ├── README.md                      # This documentation
-├── ScoringSearch.jl              # Main 23-step pipeline orchestrator
+├── PrecursorScoringSearch.jl      # Main precursor scoring pipeline orchestrator
 ├── score_psms.jl                 # PSM scoring entry point and model selection
 ├── model_config.jl               # Model configurations and feature sets
-├── utils.jl                      # Legacy protein group analysis functions
-├── utils_protein_ml.jl           # ML-enhanced protein scoring
-├── protein_inference_pipeline.jl # Modern protein inference pipeline
+├── utils.jl                      # Shared q-value / spline helpers
 └── scoring_interface.jl          # Type-safe file reference operations
 ```
 
@@ -320,7 +319,7 @@ src/utils/ML/
 ### Common Issues
 
 1. **Model Comparison Failures**: Check feature availability in PSM data
-2. **Memory Issues**: Reduce max_psms_in_memory or use out-of-memory processing
+2. **Memory Issues**: Reduce `optimization.machine_learning.max_in_memory_table_mb` or use out-of-memory processing
 3. **Poor Performance**: Inspect CV fold balance and feature distributions
 4. **Probit Convergence**: Check for sufficient training data per fold
