+++ src/Routines/SearchDIA/SearchMethods/PrecursorScoringSearch/model_config.jl
@@ -16,7 +16,7 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 #==========================================================
-Model Configuration for ScoringSearch
+Model Configuration for PrecursorScoringSearch
 This file contains only the minimal functionality needed for model selection.
 The full model_comparison.jl file is deprecated and should not be used.
 ==========================================================#
