+++ src/Routines/SearchDIA/SearchMethods/ProteinInferenceSearch/ProteinInferenceSearch.jl
@@ -1,0 +1,76 @@
+# Copyright (C) 2024 Nathan Wamsley
+#
+# This file is part of Pioneer.jl
+#
+# Pioneer.jl is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+"""
+    ProteinInferenceSearch
+
+Search method for annotating integrated passing precursor tables with inferred
+protein groups and protein-quant eligibility flags.
+"""
+struct ProteinInferenceSearch <: SearchMethod end
+
+struct ProteinInferenceSearchResults <: SearchResults end
+
+struct ProteinInferenceSearchParameters <: SearchParameters
+    function ProteinInferenceSearchParameters(::PioneerParameters)
+        new()
+    end
+end
+
+get_parameters(::ProteinInferenceSearch, params::Any) = ProteinInferenceSearchParameters(params)
+
+function init_search_results(::ProteinInferenceSearchParameters, search_context::SearchContext)
+    return ProteinInferenceSearchResults()
+end
+
+function process_file!(
+    results::ProteinInferenceSearchResults,
+    params::ProteinInferenceSearchParameters,
+    search_context::SearchContext,
+    ms_file_idx::Int64,
+    spectra::MassSpecData
+)
+    return results
+end
+
+function process_search_results!(
+    results::ProteinInferenceSearchResults,
+    params::ProteinInferenceSearchParameters,
+    search_context::SearchContext,
+    ms_file_idx::Int64,
+    spectra::MassSpecData
+)
+    return nothing
+end
+
+function reset_results!(results::ProteinInferenceSearchResults)
+    return nothing
+end
+
+function summarize_results!(
+    ::ProteinInferenceSearchResults,
+    params::ProteinInferenceSearchParameters,
+    search_context::SearchContext
+)
+    valid_file_data = get_valid_file_paths(search_context, getPassingPsms)
+    isempty(valid_file_data) && return nothing
+
+    passing_refs = [PSMFileReference(path) for (_, path) in valid_file_data]
+    run_protein_inference!(search_context; passing_refs = passing_refs)
+
+    return nothing
+end
