+++ src/Routines/SearchDIA/SearchMethods/ProteinInferenceSearch/utils.jl
@@ -1,0 +1,261 @@
+# Copyright (C) 2024 Nathan Wamsley
+#
+# This file is part of Pioneer.jl
+#
+# Pioneer.jl is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+"""
+    add_peptide_metadata(precursors::LibraryPrecursors)
+
+Add peptide sequence and protein information from the precursor library.
+"""
+function add_peptide_metadata(precursors::LibraryPrecursors)
+    desc = "add_peptide_metadata"
+
+    op = function(df)
+        all_sequences = getSequence(precursors)::AbstractVector{String}
+        all_accessions = getAccessionNumbers(precursors)::AbstractVector{String}
+        all_is_decoys = getIsDecoy(precursors)::AbstractVector{Bool}
+        all_entrap_ids = getEntrapmentGroupId(precursors)::AbstractVector{UInt8}
+        all_species = getProteomeIdentifiers(precursors)::AbstractVector{<:AbstractString}
+        all_base_pep_ids = getBasePepId(precursors)::AbstractVector{UInt32}
+        all_structural_mods = getStructuralMods(precursors)::AbstractVector{Union{Missing, String}}
+        all_isotopic_mods = getIsotopicMods(precursors)::AbstractVector{Union{Missing, String}}
+        all_pair_ids = getPairIdx(precursors)
+
+        precursor_idx = df.precursor_idx::AbstractVector{UInt32}
+        n_rows = length(precursor_idx)
+
+        if !hasproperty(df, :sequence)
+            sequences = Vector{String}(undef, n_rows)
+            for i in 1:n_rows
+                sequences[i] = all_sequences[precursor_idx[i]]
+            end
+            df.sequence = sequences
+        end
+
+        if !hasproperty(df, :accession_numbers)
+            accessions = Vector{String}(undef, n_rows)
+            for i in 1:n_rows
+                accessions[i] = all_accessions[precursor_idx[i]]
+            end
+            df.accession_numbers = accessions
+        end
+
+        if !hasproperty(df, :is_decoy)
+            is_decoy_vec = Vector{Bool}(undef, n_rows)
+            for i in 1:n_rows
+                is_decoy_vec[i] = all_is_decoys[precursor_idx[i]]
+            end
+            df.is_decoy = is_decoy_vec
+        end
+
+        if !hasproperty(df, :entrap_id)
+            entrap_vec = Vector{UInt8}(undef, n_rows)
+            for i in 1:n_rows
+                entrap_vec[i] = all_entrap_ids[precursor_idx[i]]
+            end
+            df.entrap_id = entrap_vec
+        end
+
+        if !hasproperty(df, :species)
+            species = Vector{String}(undef, n_rows)
+            for i in 1:n_rows
+                species[i] = join(sort(unique(split(coalesce(all_species[precursor_idx[i]], ""), ';'))), ';')
+            end
+            df.species = species
+        end
+
+        if !hasproperty(df, :base_pep_id)
+            base_pep_ids = Vector{UInt32}(undef, n_rows)
+            for i in 1:n_rows
+                base_pep_ids[i] = all_base_pep_ids[precursor_idx[i]]
+            end
+            df.base_pep_id = base_pep_ids
+        end
+
+        if !hasproperty(df, :structural_mods)
+            structural_mods = Vector{String}(undef, n_rows)
+            for i in 1:n_rows
+                structural_mods[i] = coalesce(all_structural_mods[precursor_idx[i]], "")
+            end
+            df.structural_mods = structural_mods
+        end
+
+        if !hasproperty(df, :isotopic_mods)
+            isotopic_mods = Vector{String}(undef, n_rows)
+            for i in 1:n_rows
+                isotopic_mods[i] = coalesce(all_isotopic_mods[precursor_idx[i]], "")
+            end
+            df.isotopic_mods = isotopic_mods
+        end
+
+        if !hasproperty(df, :pair_id)
+            pair_ids = Vector{UInt32}(undef, n_rows)
+            for i in 1:n_rows
+                pair_ids[i] = extract_pair_idx(all_pair_ids, precursor_idx[i])
+            end
+            df.pair_id = pair_ids
+        end
+
+        return df
+    end
+
+    return desc => op
+end
+
+"""
+    validate_peptide_data()
+
+Ensure all required columns exist for protein inference.
+"""
+function validate_peptide_data()
+    desc = "validate_peptide_data"
+
+    op = function(df)
+        required_cols = [
+            :sequence,
+            :accession_numbers,
+            :is_decoy,
+            :entrap_id,
+            :precursor_idx,
+            :base_pep_id,
+            :structural_mods,
+            :isotopic_mods
+        ]
+
+        missing_cols = setdiff(required_cols, Symbol.(names(df)))
+        if !isempty(missing_cols)
+            error("Missing required columns for protein inference: $missing_cols")
+        end
+
+        return df
+    end
+
+    return desc => op
+end
+
+"""
+    apply_inference_to_dataframe(df::DataFrame, precursors::LibraryPrecursors)
+
+Apply the core protein inference algorithm to a prepared DataFrame.
+"""
+function apply_inference_to_dataframe(df::DataFrame, precursors::LibraryPrecursors)
+    if nrow(df) == 0
+        return InferenceResult(Dictionary{PeptideKey, ProteinKey}())
+    end
+
+    unique_pairs = unique(df, [:sequence, :accession_numbers, :is_decoy, :entrap_id])
+    proteins_vec = Vector{ProteinKey}(undef, nrow(unique_pairs))
+    peptides_vec = Vector{PeptideKey}(undef, nrow(unique_pairs))
+
+    for (i, row) in enumerate(eachrow(unique_pairs))
+        proteins_vec[i] = ProteinKey(row.accession_numbers, !row.is_decoy, row.entrap_id)
+        peptides_vec[i] = PeptideKey(row.sequence, !row.is_decoy, row.entrap_id)
+    end
+
+    return infer_proteins(proteins_vec, peptides_vec)
+end
+
+"""
+    add_inferred_protein_column(inference_result::InferenceResult)
+
+Add inferred protein-group assignments to PSMs.
+"""
+function add_inferred_protein_column(inference_result::InferenceResult)
+    desc = "add_inferred_protein_column"
+
+    op = function(df)
+        sequences = df.sequence::AbstractVector{String}
+        is_decoy = df.is_decoy::AbstractVector{Bool}
+        entrap_ids = df.entrap_id::AbstractVector{UInt8}
+
+        inferred_proteins = Vector{Union{Missing, String}}(undef, length(sequences))
+        for i in eachindex(sequences, is_decoy, entrap_ids)
+            pep_key = PeptideKey(sequences[i], !is_decoy[i], entrap_ids[i])
+            if haskey(inference_result.peptide_to_protein, pep_key)
+                inferred_proteins[i] = inference_result.peptide_to_protein[pep_key].name
+            else
+                inferred_proteins[i] = missing
+            end
+        end
+
+        df.inferred_protein_group = inferred_proteins
+        return df
+    end
+
+    return desc => op
+end
+
+"""
+    add_quantification_flag(inference_result::InferenceResult)
+
+Mark peptides assigned to inferred protein groups as usable for protein quant/scoring.
+"""
+function add_quantification_flag(inference_result::InferenceResult)
+    desc = "add_quantification_flag"
+
+    op = function(df)
+        sequences = df.sequence::AbstractVector{String}
+        is_decoy = df.is_decoy::AbstractVector{Bool}
+        entrap_ids = df.entrap_id::AbstractVector{UInt8}
+
+        use_for_quant = Vector{Bool}(undef, length(sequences))
+        for i in eachindex(sequences, is_decoy, entrap_ids)
+            pep_key = PeptideKey(sequences[i], !is_decoy[i], entrap_ids[i])
+            use_for_quant[i] = haskey(inference_result.peptide_to_protein, pep_key)
+        end
+
+        df.use_for_protein_quant = use_for_quant
+        return df
+    end
+
+    return desc => op
+end
+
+"""
+    run_protein_inference!(search_context; passing_refs)
+
+Annotate passing precursor tables in place with inferred protein groups and
+protein-quant eligibility flags.
+"""
+function run_protein_inference!(
+    search_context::SearchContext;
+    passing_refs::Vector{PSMFileReference}
+)
+    isempty(passing_refs) && return nothing
+
+    precursors = getPrecursors(getSpecLib(search_context))
+    annotation_pipeline = TransformPipeline() |>
+        add_peptide_metadata(precursors) |>
+        validate_peptide_data()
+
+    indexed_refs = collect(enumerate(passing_refs))
+    @user_info "Annotating passing PSM files with inferred protein groups and protein-quant flags"
+
+    for (_, psm_ref) in ProgressBar(indexed_refs)
+        exists(psm_ref) || continue
+
+        apply_pipeline!(psm_ref, annotation_pipeline)
+        prepared_df = load_dataframe(psm_ref)
+        inference_result = apply_inference_to_dataframe(prepared_df, precursors)
+
+        update_pipeline = TransformPipeline() |>
+            add_inferred_protein_column(inference_result) |>
+            add_quantification_flag(inference_result)
+        apply_pipeline!(psm_ref, update_pipeline)
+    end
+
+    return nothing
+end
