+++ src/Routines/SearchDIA/SearchMethods/ProteinScoringSearch/protein_inference_pipeline.jl
@@ -37,271 +37,6 @@ end
 end
 
 """
-    add_peptide_metadata(precursors::LibraryPrecursors)
-
-Add peptide sequence and protein information from precursor library.
-Enriches PSM data with metadata needed for inference.
-"""
-function add_peptide_metadata(precursors::LibraryPrecursors)
-    desc = "add_peptide_metadata"
-    
-    op = function(df)
-        # Get library data
-        all_sequences = getSequence(precursors)::AbstractVector{String}
-        all_accessions = getAccessionNumbers(precursors)::AbstractVector{String}
-        all_is_decoys = getIsDecoy(precursors)::AbstractVector{Bool}
-        all_entrap_ids = getEntrapmentGroupId(precursors)::AbstractVector{UInt8}
-        all_species = getProteomeIdentifiers(precursors)::AbstractVector{<:AbstractString}
-        all_base_pep_ids = getBasePepId(precursors)::AbstractVector{UInt32}
-        all_structural_mods = getStructuralMods(precursors)::AbstractVector{Union{Missing, String}}
-        all_isotopic_mods = getIsotopicMods(precursors)::AbstractVector{Union{Missing, String}}
-        all_pair_ids = getPairIdx(precursors)
-        
-        # Extract precursor_idx column with type assertion for performance
-        precursor_idx = df.precursor_idx::AbstractVector{UInt32}
-        n_rows = length(precursor_idx)
-        
-        # Add columns if they don't exist
-        if !hasproperty(df, :sequence)
-            sequences = Vector{String}(undef, n_rows)
-            for i in 1:n_rows
-                sequences[i] = all_sequences[precursor_idx[i]]
-            end
-            df.sequence = sequences
-        end
-        
-        if !hasproperty(df, :accession_numbers)
-            accessions = Vector{String}(undef, n_rows)
-            for i in 1:n_rows
-                accessions[i] = all_accessions[precursor_idx[i]]
-            end
-            df.accession_numbers = accessions
-        end
-        
-        # Add target/decoy status if needed
-        if !hasproperty(df, :is_decoy)
-            is_decoy_vec = Vector{Bool}(undef, n_rows)
-            for i in 1:n_rows
-                is_decoy_vec[i] = all_is_decoys[precursor_idx[i]]
-            end
-            df.is_decoy = is_decoy_vec
-        end
-        
-        # Add entrapment group if needed
-        if !hasproperty(df, :entrap_id)
-            entrap_vec = Vector{UInt8}(undef, n_rows)
-            for i in 1:n_rows
-                entrap_vec[i] = all_entrap_ids[precursor_idx[i]]
-            end
-            df.entrap_id = entrap_vec
-        end
-
-        if !hasproperty(df, :species)
-            species = Vector{String}(undef, n_rows)
-            for i in 1:n_rows
-                species[i] = join(sort(unique(split(coalesce(all_species[precursor_idx[i]], ""), ';'))), ';')
-            end
-            df.species = species
-        end
-
-        if !hasproperty(df, :base_pep_id)
-            base_pep_ids = Vector{UInt32}(undef, n_rows)
-            for i in 1:n_rows
-                base_pep_ids[i] = all_base_pep_ids[precursor_idx[i]]
-            end
-            df.base_pep_id = base_pep_ids
-        end
-
-        if !hasproperty(df, :structural_mods)
-            structural_mods = Vector{String}(undef, n_rows)
-            for i in 1:n_rows
-                structural_mods[i] = coalesce(all_structural_mods[precursor_idx[i]], "")
-            end
-            df.structural_mods = structural_mods
-        end
-
-        if !hasproperty(df, :isotopic_mods)
-            isotopic_mods = Vector{String}(undef, n_rows)
-            for i in 1:n_rows
-                isotopic_mods[i] = coalesce(all_isotopic_mods[precursor_idx[i]], "")
-            end
-            df.isotopic_mods = isotopic_mods
-        end
-
-        if !hasproperty(df, :pair_id)
-            pair_ids = Vector{UInt32}(undef, n_rows)
-            for i in 1:n_rows
-                pair_ids[i] = extract_pair_idx(all_pair_ids, precursor_idx[i])
-            end
-            df.pair_id = pair_ids
-        end
-        
-        return df
-    end
-    
-    return desc => op
-end
-
-"""
-    validate_peptide_data()
-
-Ensure all required columns exist for protein inference.
-"""
-function validate_peptide_data()
-    desc = "validate_peptide_data"
-    
-    op = function(df)
-        required_cols = [
-            :sequence,
-            :accession_numbers,
-            :is_decoy,
-            :entrap_id,
-            :precursor_idx,
-            :base_pep_id,
-            :structural_mods,
-            :isotopic_mods
-        ]
-        
-        missing_cols = setdiff(required_cols, Symbol.(names(df)))
-        if !isempty(missing_cols)
-            error("Missing required columns for protein inference: $missing_cols")
-        end
-        
-        return df
-    end
-    
-    return desc => op
-end
-
-#==========================================================
-Inference Wrapper Functions
-==========================================================#
-
-"""
-    apply_inference_to_dataframe(df::DataFrame, precursors::LibraryPrecursors)
-
-Apply the core infer_proteins algorithm to a prepared DataFrame.
-Returns structured inference results using type-safe protein inference.
-"""
-function apply_inference_to_dataframe(df::DataFrame, precursors::LibraryPrecursors)
-    if nrow(df) == 0
-        # Return empty result for empty input
-        return InferenceResult(
-            Dictionary{PeptideKey, ProteinKey}()
-        )
-    end
-    
-    # Extract unique peptide-protein pairs
-    unique_pairs = unique(df, [:sequence, :accession_numbers, :is_decoy, :entrap_id])
-    
-    # Convert to format expected by infer_proteins
-    proteins_vec = Vector{ProteinKey}(undef, nrow(unique_pairs))
-    peptides_vec = Vector{PeptideKey}(undef, nrow(unique_pairs))
-    
-    for (i, row) in enumerate(eachrow(unique_pairs))
-        proteins_vec[i] = ProteinKey(
-            row.accession_numbers,
-            !row.is_decoy,  # Convert is_decoy to is_target
-            row.entrap_id
-        )
-        
-        peptides_vec[i] = PeptideKey(
-            row.sequence,
-            !row.is_decoy,  # Convert is_decoy to is_target
-            row.entrap_id
-        )
-    end
-    
-    # Call inference algorithm
-    return infer_proteins(proteins_vec, peptides_vec)
-end
-
-#==========================================================
-Post-Inference Pipeline Operations
-==========================================================#
-
-"""
-    add_inferred_protein_column(inference_result::InferenceResult)
-
-Add inferred protein group assignments to PSMs based on inference results.
-"""
-function add_inferred_protein_column(inference_result::InferenceResult)
-    desc = "add_inferred_protein_column"
-    
-    op = function(df)
-        # Extract columns with type assertions for performance
-        sequences = df.sequence::AbstractVector{String}
-        is_decoy = df.is_decoy::AbstractVector{Bool}
-        entrap_ids = df.entrap_id::AbstractVector{UInt8}
-        accession_numbers = df.accession_numbers::AbstractVector{String}
-        
-        n_rows = length(sequences)
-        inferred_proteins = Vector{Union{Missing, String}}(undef, n_rows)
-        
-        for i in 1:n_rows
-            pep_key = PeptideKey(
-                sequences[i],
-                !is_decoy[i],
-                entrap_ids[i]
-            )
-            
-            if haskey(inference_result.peptide_to_protein, pep_key)
-                protein_key = inference_result.peptide_to_protein[pep_key]
-                inferred_proteins[i] = protein_key.name
-            else
-                # Fallback to original protein
-                inferred_proteins[i] = missing#accession_numbers[i]
-            end
-        end
-        
-        df.inferred_protein_group = inferred_proteins
-        return df
-    end
-    
-    return desc => op
-end
-
-"""
-    add_quantification_flag(inference_result::InferenceResult)
-
-Add flag indicating whether peptide should be used for quantification.
-
-Peptides present in the inference result (unique peptides assigned to minimal protein set)
-are marked as usable for quantification. Peptides not in the result (shared peptides or
-filtered out) are marked as NOT usable for quantification.
-"""
-function add_quantification_flag(inference_result::InferenceResult)
-    desc = "add_quantification_flag"
-
-    op = function(df)
-        # Extract columns with type assertions for performance
-        sequences = df.sequence::AbstractVector{String}
-        is_decoy = df.is_decoy::AbstractVector{Bool}
-        entrap_ids = df.entrap_id::AbstractVector{UInt8}
-
-        n_rows = length(sequences)
-        use_for_quant = Vector{Bool}(undef, n_rows)
-
-        for i in 1:n_rows
-            pep_key = PeptideKey(
-                sequences[i],
-                !is_decoy[i],
-                entrap_ids[i]
-            )
-
-            # Peptide is usable for quantification if it's in the inference results
-            # (i.e., it's a unique peptide assigned to a protein)
-            use_for_quant[i] = haskey(inference_result.peptide_to_protein, pep_key)
-        end
-
-        df.use_for_protein_quant = use_for_quant
-        return df
-    end
-
-    return desc => op
-end
-
-"""
     _median_abs_deviation(values)
 
 Return the median absolute deviation of `values`.
@@ -1652,8 +1387,7 @@ function group_psms_by_protein(
             pg_score = Float32[],
             any_common_peps = Bool[],
             top_pep_weight = Float32[],
-            precursor_consensus_prefix_shape = Float32[],
-            pg_score_x_precursor_consensus_prefix_shape = Float32[]
+            precursor_consensus_prefix_shape = Float32[]
         )
     end
 
@@ -1669,8 +1403,7 @@ function group_psms_by_protein(
             pg_score = Float32[],
             any_common_peps = Bool[],
             top_pep_weight = Float32[],
-            precursor_consensus_prefix_shape = Float32[],
-            pg_score_x_precursor_consensus_prefix_shape = Float32[]
+            precursor_consensus_prefix_shape = Float32[]
         )
     end
 
@@ -1720,8 +1453,7 @@ function group_psms_by_protein(
             pg_score = pg_score,
             any_common_peps = has_common,
             top_pep_weight = top_pep_weight,
-            precursor_consensus_prefix_shape = precursor_consensus_prefix_shape,
-            pg_score_x_precursor_consensus_prefix_shape = pg_score * precursor_consensus_prefix_shape
+            precursor_consensus_prefix_shape = precursor_consensus_prefix_shape
         )
     end
 
@@ -1853,14 +1585,14 @@ High-Level Interface
 ==========================================================#
 
 """
-    perform_protein_inference_pipeline(psm_refs, output_folder, precursors, protein_catalog; kwargs...)
+    build_protein_group_tables(psm_refs, output_folder, protein_catalog; kwargs...)
 
-Perform protein inference using a composable pipeline approach.
+Build per-run protein-group tables and protein scoring features from PSM tables
+that have already been annotated with protein inference results.
 
 # Arguments
 - `psm_refs`: Vector of PSM file references
 - `output_folder`: Directory for protein group output
-- `precursors`: Library precursor information
 - `protein_catalog`: Pre-computed protein-to-peptide mappings
 - `min_peptides`: Minimum peptides per protein group (default: 2)
 
@@ -1868,10 +1600,9 @@ Perform protein inference using a composable pipeline approach.
 - `pg_refs`: Vector of protein group file references
 - `psm_to_pg_mapping`: Dictionary mapping PSM paths to protein group paths
 """
-function perform_protein_inference_pipeline(
+function build_protein_group_tables(
     psm_refs::Vector{PSMFileReference},
     output_folder::String,
-    precursors::LibraryPrecursors,
     protein_catalog::Dict;
     min_peptides::Int = 2,
     q_value_threshold::Float32 = 0.01f0,
@@ -1881,30 +1612,9 @@ function perform_protein_inference_pipeline(
 )
     !isdir(output_folder) && mkpath(output_folder)
 
-    pre_inference_pipeline = TransformPipeline() |>
-        add_peptide_metadata(precursors) |>
-        validate_peptide_data()
-
     pg_refs = ProteinGroupFileReference[]
     psm_to_pg_mapping = Dict{String, String}()
     indexed_refs = collect(enumerate(psm_refs))
-
-    @user_info "Annotating passing PSM files with inferred protein groups and protein-quant flags"
-
-    for (idx, psm_ref) in ProgressBar(indexed_refs)
-        if !exists(psm_ref)
-            continue
-        end
-
-        apply_pipeline!(psm_ref, pre_inference_pipeline)
-        prepared_df = load_dataframe(psm_ref)
-        inference_result = apply_inference_to_dataframe(prepared_df, precursors)
-
-        psm_update_pipeline = TransformPipeline() |>
-            add_inferred_protein_column(inference_result) |>
-            add_quantification_flag(inference_result)
-        apply_pipeline!(psm_ref, psm_update_pipeline)
-    end
 
     precursor_consensus = build_precursor_consensus(psm_refs; q_value_threshold = q_value_threshold)
     consensus_debug_rows = NamedTuple[]
