+++ src/structs/protein_inference_types.jl
@@ -16,7 +16,7 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 """
-Data types for protein inference and scoring in the ScoringSearch module.
+Data types for protein inference and scoring in the PrecursorScoringSearch module.
 
 These types provide clear, type-safe structures for protein group analysis,
 replacing complex NamedTuple-based dictionaries.
